pyproject.toml
[project]
name = "authentic"
version = "0.1.0"
description = "AI Agent/MCP 인증 서버"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "httpx>=0.26.0",
    "authlib>=1.3.0",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "itsdangerous>=2.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
